Working on communication templates files refactoring
