+++ public/views/agora-communication-templates-common.php
